feat: add validation on delete catalog

+++ src/modules/catalog/catalog.service.ts
@@ -1,12 +1,23 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  ConflictException,
+  Inject,
+  Injectable,
+  forwardRef,
+} from '@nestjs/common';
 import { CatalogRepository } from './catalog.repository';
 import { CreateCatalogDto } from './dto/create-catalog.dto';
 import { UpdateCatalogDto } from './dto/update-catalog.dto';
 import { PaginationOptions } from 'src/common/interfaces/base.interface';
+import { ProductService } from '../product/product.service';
 
 @Injectable()
 export class CatalogService {
-  constructor(private readonly catalogRepository: CatalogRepository) {}
+  constructor(
+    private readonly catalogRepository: CatalogRepository,
+    @Inject(forwardRef(() => ProductService))
+    private readonly productService: ProductService,
+  ) {}
 
   async create(data: CreateCatalogDto) {
     const existingCatalog = await this.catalogRepository.findByName(data.name);
@@ -37,6 +48,13 @@ export class CatalogService {
   }
 
   async delete(id: number) {
+    const products = await this.productService.findAllProducts({ filters: { catalogId: id } });
+
+    if (products.length > 0) {
+      throw new BadRequestException(
+        `Cannot delete catalog with ID ${id}. It has ${products.length} associated products.`,
+      );
+    }
     return this.catalogRepository.delete(id);
   }
 }
